Switch migration version from int to timestamp (bigint)  #87

+++ src/DbLive/Adapter/MigrationItemSaveDto.cs
@@ -3,7 +3,7 @@
 [ExcludeFromCodeCoverage]
 public class MigrationItemSaveDto
 {
-	public required int Version { get; init; }
+	public required long Version { get; init; }
 	public required string Name { get; init; }
 	public required string RelativePath { get; init; }
 	public required MigrationItemType ItemType { get; init; }
