+++ src/DbLive/Adapter/MigrationItemStateDto.cs
@@ -3,7 +3,7 @@
 [ExcludeFromCodeCoverage]
 public class MigrationItemStateDto
 {
-	public required int Version { get; init; }
+	public required long Version { get; init; }
 	public required MigrationItemType ItemType { get; init; }
 	public required MigrationItemStatus Status { get; init; }
 	public DateTime? AppliedUtc { get; set; }
