+++ src/DbLive/SelfDeployer/InternalMigration.cs
@@ -3,7 +3,7 @@ namespace DbLive.SelfDeployer;
 
 public class InternalMigration
 {
-	public required int Version { get; set; }
+	public required long Version { get; set; }
 	public required FileData FileData { get; set; }
 	public string Name { get; set; } = string.Empty;
 }
